refactor(chat): neutralize goal and chat utility service paths
