+++ turbo/apps/api/src/signals/services/zero-goal-objective-brief.service.ts
@@ -5,7 +5,7 @@ import {
   capGoalObjectiveBriefText,
   compactGoalObjectiveBriefText,
   fallbackGoalObjectiveBrief,
-} from "./zero-goal-objective-brief-normalization.service";
+} from "./goal-objective-brief-normalization.service";
 
 const log = logger("api:zero-goal-objective-brief");
 const OBJECTIVE_BRIEF_MODEL = "google/gemini-3.1-flash-lite-preview";
