+++ turbo/apps/api/src/signals/services/zero-goal-queue-drain.service.ts
@@ -22,7 +22,7 @@ import {
 } from "./chat-goal-queue.service";
 import type { InternalRunCallbackKind } from "./internal-run-callback";
 import { resolveRunChatThreadModelContext } from "./zero-chat-run-event.service";
-import { normalizeGoalObjectiveBrief } from "./zero-goal-objective-brief-normalization.service";
+import { normalizeGoalObjectiveBrief } from "./goal-objective-brief-normalization.service";
 import type { ModelFirstPin } from "./zero-model-selection.service";
 import { createQueueFirstZeroRun$ } from "./zero-runs-create.service";
