+++ turbo/apps/api/src/signals/services/zero-workflow-automation-launch.service.ts
@@ -23,7 +23,7 @@ import {
 } from "./api-dispatch-timing.service";
 import { createQueueFirstZeroRun$ } from "./zero-runs-create.service";
 import { workflowAutomationCanFire } from "./zero-workflow-automation-access.service";
-import { loadComputerUseHostGrantForAutoSend } from "./zero-chat-computer-use-host.service";
+import { loadComputerUseHostGrantForAutoSend } from "./chat-computer-use-host.service";
 import type { WorkflowAutomationContext } from "./workflow-automation-context.service";
 import type { ChatAgentRunSourceAnnotation } from "./zero-chat-user-message.service";
